http gw: fix races in the areas
