+++ ydb/core/kqp/proxy_service/ut/ya.make
@@ -19,6 +19,7 @@ PEERDIR(
     ydb/core/kqp/run_script_actor
     ydb/core/kqp/proxy_service
     ydb/core/kqp/ut/common
+    ydb/library/yql/providers/common/http_gateway/ut_helpers
     ydb/services/workload_manager/ut/common
     ydb/public/lib/ut_helpers
     ydb/public/sdk/cpp/src/client/driver
