+++ ydb/services/workload_manager/ut/common/ya.make
@@ -8,6 +8,7 @@ SRCS(
 PEERDIR(
     ydb/core/base
     ydb/core/kqp/ut/common
+    ydb/library/yql/providers/common/http_gateway/ut_helpers
     ydb/services/metadata
     ydb/services/workload_manager/metadata_subscription/resource_pool_classifier
 )
